game 4 is invidiual file
