+++ lara_games_routes/lara_games_routes.py
@@ -271,47 +271,10 @@ def loading_file_pic_g3(filename,pic_name):
 '''
 """GAME 4"""
 
-
+from games_logic.game4 import generate_game4 as game4
 @app.route('/game4/<story_name>', methods=['GET'])
-def generate_game4(story_name, file=None):
-    # TODO add template game 4
-    # TODO add redirecter to game4
-    # TODO print build dic from metadata file
-    # TODO get the metadata folder, random of all of these that inside the folder
-
-    metaDataAudioDir = mypath + slash_clean + story_name + slash_clean + 'audio' + slash_clean
-    audioVersions = dirinDir(metaDataAudioDir)
-    # TODO add expectaion
-    File = metaDataAudioDir + audioVersions[0] + slash_clean + 'metadata_help.json'
-    Metadata = lara_utils.read_json_file(File)
-    meta_dic = [{}]
-    for m in Metadata:
-        meta_dic[0].update({m['text']: m['file']})
-    sentance = []
-    sounds = []
-    for key, value in meta_dic[0].items():
-        sentance.append(key)
-        sounds.append(value)
-
-    """get random sentance"""
-    size_of_story = len(sentance)
-    rand_index = random.randint(0, size_of_story)
-
-    """gather 4 random index for 4 wrong answers """
-    rand_i = rand_index
-    fake_answer = []
-    for i in range(4):
-        rand_i = random.sample(range(0, size_of_story), 4)
-
-    true_match = [sentance[rand_index], sounds[rand_index]]
-    bad_match = [sentance[rand_i[0]], sentance[rand_i[1]], sentance[rand_i[2]], sentance[rand_i[3]]]
-
-    print(true_match)
-    print(bad_match)
-
-    return render_template('game4_template.html', t_answer=true_match[0], question=true_match[1],
-                           fake_answer_0=bad_match[0], fake_answer_1=bad_match[1], fake_answer_2=bad_match[2],
-                           fake_answer_3=bad_match[3], name=story_name)
+def generate_game4(story_name):
+    return game4(story_name)
     # return render_template('game4_template.html',meta_dic=meta_dic,alphabet=alphaBet,name=story_name)
 
 
@@ -323,27 +286,10 @@ def loading_file_pic_g4(filename, story_name):
     # File = metaDataAudioDir + audioVersions[0] + '/'+filename
     return send_from_directory(metaDataAudioDir + slash_clean + audioVersions[0], filename)
 
-
+from games_logic.game4 import submit_g4 as submit
 @app.route('/game4Submit/', methods=['GET', 'POST'])
 def submit_g4(option=0,answer=0,default_value=0):
-    option = request.form.get('option')
-
-    name = request.form.get('storyname', default_value)
-    answer = request.form.get('answer', default_value)
-    uid = request.form.get('uid', default_value)
-    print("name: ", name)
-    print("option: ", option)
-    print("answer: ", answer)
-
-    if option.lower() == answer.lower():
-        item = tbl_game4(score=1, user_id=uid,question=name)
-        flash('Right answer', 'success')
-    else:
-        item = tbl_game4(score=0, user_id=uid,question=name)
-        flash('bad answer', 'danger')
-    db.session.add(item)
-    db.session.commit()
-    return redirect(url_for('app.generate_game4', story_name=name))
+   return submit(option, answer, default_value)
 
 
 """GAME 5"""
